+++ .env.example
@@ -1,0 +1,1 @@
+VITE_NYTIMES_API_KEY=
