+++ src/components/ArtsFeedComponent/index.tsx
@@ -1,3 +1,4 @@
+import { fetchArtsFeed } from '@local/services/NYTimesAPI';
 import { useEffect, useState } from 'react';
 
 export default function ArtsFeedComponent(): JSX.Element {
@@ -5,43 +6,17 @@ export default function ArtsFeedComponent(): JSX.Element {
 	const [error, setError] = useState<string | null>(null);
 
 	useEffect(() => {
-		const fetchArtsFeed = async () => {
+		const getArtsFeed = async () => {
 			try {
-				const response = await fetch(
-					'https://rss.nytimes.com/services/xml/rss/nyt/Arts.xml'
-				);
-
-				if (!response.ok) {
-					throw new Error('Failed to fetch arts feed');
-				}
-
-				const data = await response.text();
-				const parser = new DOMParser();
-				const xmlDocument = parser.parseFromString(data, 'text/xml');
-				const items = xmlDocument.querySelectorAll('item');
-
-				const articlesData = Array.from(items).map((item) => {
-					return {
-						title: item.querySelector('title')?.textContent,
-						link: item.querySelector('link')?.textContent,
-						description:
-							item.querySelector('description')?.textContent,
-						pubDate: item.querySelector('pubDate')?.textContent,
-						category: item.querySelector('category')?.textContent,
-						image: item
-							.querySelector('media\\:content, [url]')
-							?.getAttribute('url'), // Adjust this selector based on the structure of the feed
-					};
-				});
-
-				setArticles(articlesData);
+				const data = await fetchArtsFeed();
+				setArticles(data);
 			} catch (error: any) {
 				setError(error.message);
 				console.error(error);
 			}
 		};
 
-		fetchArtsFeed();
+		getArtsFeed();
 	}, []);
 
 	return (
